+++ CHANGELOG.md
@@ -20,7 +20,9 @@ Fleam release notes
 * Adds the option to process the SQS delete `BatchResult`s via a flow passed to the `SqsStreamDaemon`.
 * Changes `MissingMessageGroupId` to `MissingGroupingKey` for `SqsReduce` to make it more general.
 * Adds value class for `MessageGroupId`. Usages of SqsReduce will need to switch from `String` to `MessageGroupId`.
+* Change the default config value for SQS long polling from 0 to 20 seconds
 * Changes SerializedByKeyBidi to take a buffer instead of forcing one.
+
 
 ## Fleam 6.0.0
 * Adds the ability to recover exceptions while using a Valve to avoid having elements dropped by akka streams